+++ changelog.md
@@ -3,7 +3,10 @@ Naia - Project Ethea Phase 1 Shared Library - Changelog
 
 Version 20200814+dev:
 ---------------------
-
+* Lua and WML library:
+  * Fixed persistent WML variables API crashing when trying to manipulate the
+    global table on a fresh persistent store or a store without a preexisting
+    global table.
 
 
 Version 20200814: